Removed Print Statementws
I was curious as to just how many prefab get loaded into the pool (Queue).  It turned out that  the Count of the object pool never went above one, which was a little surprising at first, but made more sense as I thought about it.  1 goes in,, 1 comes out.  Still--I'm not quite convinced that this isn't situation dependent.  If I preloaded the Pool with 10 objects, I think it would fall back to 1, but I'm not positieve.